requirements.txt
pytest==9.0.2
python-dotenv==1.2.1
watchdog==6.0.0
ruff==0.14.14
